refactor #27 : reorganize ErrorCode classes into errorcode package
- Move all ErrorCode enums to com.phonebid.app.common.errorcode package
- Update import statements across all affected files
- Improve code organization with clear separation of concerns
