+++ backend/src/main/java/com/phonebid/app/common/exception/CustomException.java
@@ -1,5 +1,6 @@
 package com.phonebid.app.common.exception;
 
+import com.phonebid.app.common.errorcode.ErrorCode;
 import lombok.Getter;
 
 /**
