+++ backend/src/main/java/com/phonebid/app/member/service/NaverService.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.phonebid.app.common.exception.CustomException;
-import com.phonebid.app.common.exception.NaverErrorCode;
-import com.phonebid.app.common.exception.CommonErrorCode;
+import com.phonebid.app.common.errorcode.NaverErrorCode;
+import com.phonebid.app.common.errorcode.CommonErrorCode;
 import com.phonebid.app.jwt.JwtUtil;
 import com.phonebid.app.member.domain.Provider;
 import com.phonebid.app.member.domain.Role;
