+++ backend/src/main/java/com/phonebid/app/member/service/UserService.java
@@ -3,8 +3,8 @@ package com.phonebid.app.member.service;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import com.phonebid.app.common.exception.CommonErrorCode;
 import com.phonebid.app.common.exception.CustomException;
+import com.phonebid.app.common.errorcode.CommonErrorCode;
 import com.phonebid.app.member.dto.RequestDto.SignupRequestDto;
 import com.phonebid.app.member.dto.RequestDto.LoginRequestDto;
 import com.phonebid.app.member.dto.ResponseDto.LoginResponseDto;
